+++ refinery/data_set_manager/tests.py
@@ -22,12 +22,11 @@ from celery.states import FAILURE, PENDING, STARTED, SUCCESS
 from djcelery.models import TaskMeta
 from factory_boy.utils import (create_dataset_with_necessary_models,
                                make_analyses_with_single_dataset)
-from guardian.shortcuts import assign_perm
 from haystack.exceptions import SkipDocument
 import mock
 from mock import ANY
 from override_storage import override_storage
-from rest_framework.test import APIClient, APIRequestFactory, APITestCase
+from rest_framework.test import APITestCase
 
 import constants
 from core.models import (INPUT_CONNECTION, OUTPUT_CONNECTION, Analysis,
@@ -53,476 +52,10 @@ from .utils import (_create_solr_params_from_node_uuids,
                     hide_fields_from_list, initialize_attribute_order_ranks,
                     insert_facet_field_filter, is_field_in_hidden_list,
                     objectify_facet_field_counts, update_attribute_order_ranks)
-from .views import Assays, AssaysAttributes
 
 TEST_DATA_BASE_PATH = "data_set_manager/test-data/"
 
 logger = logging.getLogger(__name__)
-
-
-class AssaysAPITests(APITestCase):
-
-    def setUp(self):
-        self.factory = APIRequestFactory()
-        investigation = Investigation.objects.create()
-        self.study = Study.objects.create(
-                file_name='test_filename123.txt',
-                title='Study Title Test',
-                investigation=investigation)
-        self.assay = {
-            'study': self.study,
-            'measurement': 'transcription factor binding site',
-            'measurement_accession': 'http://www.testurl.org/testID',
-            'measurement_source': 'OBI',
-            'technology': 'nucleotide sequencing',
-            'technology_accession': 'test info',
-            'technology_source': 'test source',
-            'platform': 'Genome Analyzer II',
-            'file_name': 'test_assay_filename.txt'
-        }
-        assay = Assay.objects.create(**self.assay)
-        self.assay['uuid'] = assay.uuid
-        self.assay['study'] = self.study.id
-        self.valid_uuid = assay.uuid
-        self.url_root = '/api/v2/assays/'
-        self.view = Assays.as_view()
-        self.invalid_uuid = "0xxx000x-00xx-000x-xx00-x00x00x00x0x"
-        self.invalid_format_uuid = "xxxxxxxx"
-
-    def test_get_valid_uuid(self):
-        # valid_uuid
-        request = self.factory.get('%s/?uuid=%s' % (
-            self.url_root, self.valid_uuid))
-        response = self.view(request, self.valid_uuid)
-        self.assertEqual(response.status_code, 200)
-        self.assertItemsEqual(response.data.keys(), self.assay.keys())
-        self.assertItemsEqual(response.data.values(), self.assay.values())
-
-    def test_get_valid_study(self):
-        # valid_study_uuid
-        request = self.factory.get('%s/?study=%s' % (
-            self.url_root, self.study.uuid))
-        response = self.view(request, self.valid_uuid)
-        self.assertEqual(response.status_code, 200)
-        self.assertItemsEqual(response.data[0].keys(), self.assay.keys())
-        self.assertItemsEqual(response.data[0].values(), self.assay.values())
-
-    def test_get_invalid_uuid(self):
-        # invalid_uuid
-        request = self.factory.get('%s/?uuid=%s' % (self.url_root,
-                                                    self.invalid_uuid))
-        response = self.view(request, self.invalid_uuid)
-        self.assertEqual(response.status_code, 404)
-
-    def test_get_invalid_study_uuid(self):
-        # invalid_study_uuid
-        request = self.factory.get('%s/?study=%s' % (self.url_root,
-                                                     self.invalid_uuid))
-        response = self.view(request, self.invalid_uuid)
-        self.assertEqual(response.status_code, 404)
-
-    def test_get_invalid_format_uuid(self):
-        # invalid_format_uuid
-        request = self.factory.get('%s/?uuid=%s'
-                                   % (self.url_root, self.invalid_format_uuid))
-        response = self.view(request, self.invalid_format_uuid)
-        self.assertEqual(response.status_code, 404)
-
-
-class AssaysAttributesAPITests(APITestCase):
-
-    def setUp(self):
-        self.user1 = User.objects.create_user("ownerJane", '', 'test1234')
-        self.user2 = User.objects.create_user("guestName", '', 'test1234')
-        self.factory = APIRequestFactory()
-        investigation = Investigation.objects.create()
-        self.data_set = DataSet.objects.create(
-                title="Test DataSet")
-        InvestigationLink.objects.create(data_set=self.data_set,
-                                         investigation=investigation)
-        self.data_set.set_owner(self.user1)
-        study = Study.objects.create(file_name='test_filename123.txt',
-                                     title='Study Title Test',
-                                     investigation=investigation)
-
-        assay = Assay.objects.create(
-                study=study,
-                measurement='transcription factor binding site',
-                measurement_accession='http://www.testurl.org/testID',
-                measurement_source='OBI',
-                technology='nucleotide sequencing',
-                technology_accession='test info',
-                technology_source='test source',
-                platform='Genome Analyzer II',
-                file_name='test_assay_filename.txt')
-
-        self.attribute_order_array = [
-            {
-                'study': study,
-                'assay': assay,
-                'solr_field': 'Character_Title_6_3_s',
-                'rank': 1,
-                'is_exposed': True,
-                'is_facet': True,
-                'is_active': True,
-                'is_internal': False
-            }, {
-                'study': study,
-                'assay': assay,
-                'solr_field': 'Specimen_6_3_s',
-                'rank': 2,
-                'is_exposed': True,
-                'is_facet': True,
-                'is_active': True,
-                'is_internal': False
-            }, {
-                'study': study,
-                'assay': assay,
-                'solr_field': 'Cell_Type_6_3_s',
-                'rank': 3,
-                'is_exposed': True,
-                'is_facet': True,
-                'is_active': True,
-                'is_internal': False
-            }, {
-                'study': study,
-                'assay': assay,
-                'solr_field': 'Analysis_6_3_s',
-                'rank': 4,
-                'is_exposed': True,
-                'is_facet': True,
-                'is_active': True,
-                'is_internal': False
-            }]
-
-        self.attribute_order_response = [
-            {
-                'study': study,
-                'assay': assay,
-                'solr_field': 'Character_Title_6_3_s',
-                'rank': 1,
-                'is_exposed': True,
-                'is_facet': True,
-                'is_active': True,
-                'is_internal': False,
-                'display_name': 'Character Title'
-            }, {
-                'study': study,
-                'assay': assay,
-                'solr_field': 'Specimen_6_3_s',
-                'rank': 2,
-                'is_exposed': True,
-                'is_facet': True,
-                'is_active': True,
-                'is_internal': False,
-                'display_name': 'Specimen'
-            }, {
-                'study': study,
-                'assay': assay,
-                'solr_field': 'Cell_Type_6_3_s',
-                'rank': 3,
-                'is_exposed': True,
-                'is_facet': True,
-                'is_active': True,
-                'is_internal': False,
-                'display_name': 'Cell Type'
-            }, {
-                'study': study,
-                'assay': assay,
-                'solr_field': 'Analysis_6_3_s',
-                'rank': 4,
-                'is_exposed': True,
-                'is_facet': True,
-                'is_active': True,
-                'is_internal': False,
-                'display_name': 'Analysis'
-            }]
-
-        index = 0
-        for attribute in self.attribute_order_array:
-            response = AttributeOrder.objects.create(**attribute)
-            attribute['id'] = response.id
-            attribute['assay'] = response.assay.id
-            attribute['study'] = response.study.id
-            self.attribute_order_response[index]['id'] = response.id
-            self.attribute_order_response[index]['assay'] = response.assay.id
-            self.attribute_order_response[index]['study'] = response.study.id
-            index = index + 1
-
-        list.sort(self.attribute_order_response)
-        self.valid_uuid = assay.uuid
-        self.url_root = '/api/v2/assays'
-        self.view = AssaysAttributes.as_view()
-        self.invalid_uuid = "0xxx000x-00xx-000x-xx00-x00x00x00x0x"
-        self.invalid_format_uuid = "xxxxxxxx"
-
-    def test_get_valid_uuid(self):
-        # valid_uuid
-        uuid = self.valid_uuid
-        request = self.factory.get('%s/%s/attributes/' % (self.url_root, uuid))
-        response = self.view(request, uuid)
-        self.assertEqual(response.status_code, 200)
-
-        list.sort(response.data)
-
-        ind = 0
-        for attributes in response.data:
-            self.assertItemsEqual(
-                    self.attribute_order_response[ind].keys(),
-                    attributes.keys())
-            self.assertItemsEqual(
-                    self.attribute_order_response[ind].values(),
-                    attributes.values())
-            ind = ind + 1
-
-    def test_get_invalid_uuid(self):
-        # invalid uuid
-        request = self.factory.get('%s/%s/attributes/' % (
-            self.url_root, self.invalid_uuid))
-        response = self.view(request, self.invalid_uuid)
-        response.render()
-        self.assertEqual(response.status_code, 404)
-        self.assertEqual(response.content, '{"detail":"Not found."}')
-
-    def test_get_invalid_form_uuid(self):
-        # invalid form uuid
-        request = self.factory.get('%s/%s/attributes/' % (
-            self.url_root, self.invalid_format_uuid))
-        response = self.view(request, self.invalid_format_uuid)
-        response.render()
-        self.assertEqual(response.status_code, 404)
-        self.assertEqual(response.content, '{"detail":"Not found."}')
-
-    def test_put_valid_uuid(self):
-        # valid_uuid
-        self.client.login(username='ownerJane', password='test1234')
-        updated_attribute_1 = {'solr_field': 'Character_Title_6_3_s',
-                               'rank': 3,
-                               'is_exposed': False,
-                               'is_facet': False,
-                               'is_active': False}
-        id = self.attribute_order_array[2].get('id')
-        updated_attribute_2 = {'id': id,
-                               'rank': 1,
-                               'is_exposed': False,
-                               'is_facet': False,
-                               'is_active': False}
-        # Api client needs url to end / or it will redirect
-        # update with solr_title
-        response = self.client.put(
-                '{0}/{1}/attributes/'.format(
-                        self.url_root, self.valid_uuid), updated_attribute_1)
-        self.assertEqual(response.status_code, 202)
-        self.assertEqual(response.data.get('rank'), updated_attribute_1.get(
-                'rank'))
-        self.assertEqual(
-                response.data.get('is_exposed'),
-                updated_attribute_1.get('is_exposed'))
-        self.assertEqual(
-                response.data.get('is_facet'),
-                updated_attribute_1.get('is_facet'))
-
-        # Update with attribute_order id
-        response = self.client.put(
-                '{0}/{1}/attributes/'.format(
-                        self.url_root, self.valid_uuid), updated_attribute_2)
-        self.assertEqual(response.status_code, 202)
-        self.assertEqual(response.data.get('rank'),
-                         updated_attribute_2.get('rank'))
-        self.assertEqual(
-                response.data.get('is_exposed'),
-                updated_attribute_2.get('is_exposed'))
-        self.assertEqual(
-                response.data.get('is_facet'),
-                updated_attribute_2.get('is_facet'))
-        self.client.logout()
-
-    def test_put_invalid_object(self):
-        # Invalid objects
-        updated_attribute_3 = {'rank': '4',
-                               'is_exposed': 'False',
-                               'is_facet': 'False',
-                               'is_active': 'False'}
-
-        self.client.login(username='ownerJane', password='test1234')
-        response = self.client.put('{0}/{1}/attributes/'
-                                   .format(self.url_root, self.valid_uuid),
-                                   updated_attribute_3)
-        response.render()
-        self.assertEqual(response.status_code, 400)
-        self.assertEqual(
-                response.content, '"Requires attribute id or solr_field name."'
-                )
-        self.client.logout()
-
-    def test_put_invalid_login(self):
-        # Invalid Login
-        updated_attribute_4 = {'solr_field': 'Cell Type',
-                               'rank': '4',
-                               'is_exposed': 'False',
-                               'is_facet': 'False',
-                               'is_active': 'False'}
-
-        self.client.login(username='guestName', password='test1234')
-        response = self.client.put('{0}/{1}/attributes/'
-                                   .format(self.url_root, self.valid_uuid),
-                                   updated_attribute_4)
-        self.assertEqual(response.status_code, 401)
-        self.assertEqual(
-                response.content, '"Only owner may edit attribute order."'
-                )
-        self.client.logout()
-
-
-class AssaysFilesAPITests(APITestCase):
-
-    def setUp(self):
-        self.user_owner = 'owner'
-        self.user_guest = 'guest'
-        self.fake_password = 'test1234'
-        self.data_set = create_dataset_with_necessary_models()
-        self.user1 = User.objects.create_user(self.user_owner,
-                                              '',
-                                              self.fake_password)
-        self.user2 = User.objects.create_user(self.user_guest,
-                                              '',
-                                              self.fake_password)
-        self.data_set.set_owner(self.user1)
-        investigation = Investigation.objects.create()
-        self.investigation_link = \
-            InvestigationLink.objects.create(investigation=investigation,
-                                             data_set=self.data_set,
-                                             version=1)
-
-        study = Study.objects.create(file_name='test_filename123.txt',
-                                     title='Study Title Test',
-                                     investigation=investigation)
-
-        assay = Assay.objects.create(
-                study=study,
-                measurement='transcription factor binding site',
-                measurement_accession='http://www.testurl.org/testID',
-                measurement_source='OBI',
-                technology='nucleotide sequencing',
-                technology_accession='test info',
-                technology_source='test source',
-                platform='Genome Analyzer II',
-                file_name='test_assay_filename.txt',
-                )
-        self.valid_uuid = assay.uuid
-        self.invalid_uuid = "0xxx000x-00xx-000x-xx00-x00x00x00x0x"
-        self.url = "/api/v2/assays/%s/files/"
-        self.non_meta_attributes = ['REFINERY_DOWNLOAD_URL', 'REFINERY_NAME']
-        self.client = APIClient()
-
-    def tearDown(self):
-        self.client.logout()
-        super(AssaysFilesAPITests, self).tearDown()
-
-    @mock.patch('data_set_manager.views.generate_solr_params_for_assay')
-    @mock.patch('data_set_manager.views.search_solr')
-    @mock.patch('data_set_manager.views.format_solr_response')
-    def test_get_from_owner_with_valid_params(self,
-                                              mock_format,
-                                              mock_search,
-                                              mock_generate):
-        self.client.login(username=self.user_owner,
-                          password=self.fake_password)
-        mock_format.return_value = {'status': 200}
-        uuid = self.valid_uuid
-        params = {
-            'limit': '0',
-            'data_set_uuid': self.data_set.uuid
-        }
-        response = self.client.get(self.url % uuid, params)
-        self.assertTrue(mock_format.called)
-        self.assertTrue(mock_search.called)
-        qdict = QueryDict('', mutable=True)
-        qdict.update(params)
-        mock_generate.assert_called_once_with(qdict, uuid)
-        self.assertEqual(response.status_code, 200)
-
-    def test_get_from_owner_invalid_params(self):
-        self.client.login(username=self.user_owner,
-                          password=self.fake_password)
-
-        uuid = self.valid_uuid
-        params = {'limit': 0,
-                  'data_set_uuid': self.invalid_uuid}
-        response = self.client.get(self.url % uuid, params)
-        self.assertEqual(response.status_code, 404)
-
-    @mock.patch('data_set_manager.views.generate_solr_params_for_assay')
-    @mock.patch('data_set_manager.views.search_solr')
-    @mock.patch('data_set_manager.views.format_solr_response')
-    def test_get_from_user_no_perms(self,
-                                    mock_format,
-                                    mock_search,
-                                    mock_generate):
-        self.client.login(username=self.user_guest,
-                          password=self.fake_password)
-
-        uuid = self.valid_uuid
-        params = {
-            'limit': 0,
-            'data_set_uuid': self.data_set.uuid
-        }
-        response = self.client.get(self.url % uuid, params)
-        self.assertFalse(mock_format.called)
-        self.assertFalse(mock_search.called)
-        self.assertFalse(mock_generate.called)
-        self.assertEqual(response.status_code, 401)
-
-    @mock.patch('data_set_manager.views.generate_solr_params_for_assay')
-    @mock.patch('data_set_manager.views.search_solr')
-    @mock.patch('data_set_manager.views.format_solr_response')
-    def test_get_from_user_with_read_perms(self,
-                                           mock_format,
-                                           mock_search,
-                                           mock_generate):
-        mock_format.return_value = {'status': 200}
-        self.client.login(username=self.user_guest,
-                          password=self.fake_password)
-        assign_perm('read_%s' % DataSet._meta.model_name,
-                    self.user2,
-                    self.data_set)
-        uuid = self.valid_uuid
-        params = {'limit': '0',
-                  'data_set_uuid': self.data_set.uuid}
-        response = self.client.get(self.url % uuid, params)
-        self.assertTrue(mock_format.called)
-        self.assertTrue(mock_search.called)
-        qdict = QueryDict('', mutable=True)
-        qdict.update(params)
-        mock_generate.assert_called_once_with(qdict, uuid)
-        self.assertEqual(response.status_code, 200)
-
-    @mock.patch('data_set_manager.views.generate_solr_params_for_assay')
-    @mock.patch('data_set_manager.views.search_solr')
-    @mock.patch('data_set_manager.views.format_solr_response')
-    def test_get_from_user_with_read_meta_perms(self,
-                                                mock_format,
-                                                mock_search,
-                                                mock_generate):
-        mock_format.return_value = {'status': 200}
-        self.client.login(username=self.user_guest,
-                          password=self.fake_password)
-        assign_perm('read_meta_%s' % DataSet._meta.model_name,
-                    self.user2,
-                    self.data_set)
-
-        uuid = self.valid_uuid
-        params = {'limit': '0',
-                  'data_set_uuid': self.data_set.uuid}
-        response = self.client.get(self.url % uuid, params)
-        self.assertTrue(mock_format.called)
-        self.assertTrue(mock_search.called)
-        qdict = QueryDict('', mutable=True)
-        qdict.update(params)
-        mock_generate.assert_called_once_with(qdict,
-                                              uuid,
-                                              self.non_meta_attributes)
-        self.assertEqual(response.status_code, 200)
 
 
 class UtilitiesTests(TestCase):
